Reformatted the folder structure of the codebase
