+++ frontend/learno/src/pages/Mentee/MenteeChat.jsx
@@ -1,34 +1,36 @@
 import React, { useEffect, useState } from 'react'
-import ChatRoom from './ChatRoom'
+import WebSocketClient from '../../components/websocket/ChatRoom'
+import ChatRoom from '../../components/websocket/ChatRoom'
 import axios from 'axios';
-import { API_URL, getUserId } from '../config/config';
-import { useParams } from 'react-router-dom';
+import { API_URL, getUserId } from '../../config/config';
 import { useSelector } from 'react-redux';
+import { useParams } from 'react-router-dom';
 
-const MentorChat = () => {
+const MenteeChat = () => {
     const [mentee, setMentee] = useState(null);
     const userId = getUserId(localStorage.getItem("token"));
-    let params = useParams();
+    const currentUser = useSelector((state) => state.user.user);
+     let params = useParams();
     const receiverId = params.receiverId;
-    const currentUser = useSelector((state) => state.user.user)
 
     // useEffect(()=> {
     //     const getUser = async() => {
     //         try {
-    //             const response = await axios.get(`${API_URL}/auth/getUsers/${userId}`)
+    //             const response = await axios.get(`${API_URL}/auth/getUsers/${user.id}`)
     //             setCurrentUser(response.data)
+    //             console.log(response.data)
 
     //         } catch (err) {
     //             console.log(err.message)
     //         }
     //     }
     //     getUser();
-    // }, [userId])
+    // }, [user.id])
 
      useEffect(()=> {
         const getMentee = async() => {
             try {
-                const response = await axios.get(`${API_URL}/auth/getMentor/${userId}`,{})
+                const response = await axios.get(`${API_URL}/auth/getMentee/${userId}`,{})
                 setMentee(response.data)
 
             } catch (err) {
@@ -41,11 +43,7 @@ const MentorChat = () => {
   return (
     <div>
          {currentUser ? (
-            <ChatRoom 
-                    currentUser={currentUser} 
-                    roleDetails={mentee} 
-                    receiverId={receiverId} 
-                />
+            <ChatRoom currentUser={currentUser} roleDetails={mentee} receiverId={receiverId} />
             ) : (
             <p>Loading chat...</p> 
             )}
@@ -53,4 +51,4 @@ const MentorChat = () => {
   )
 }
 
-export default MentorChat
+export default MenteeChat
